# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API Documentation using Swagger"
  },
  "paths": {
    "/adddeviceinfo": {
      "post": {
        "summary": "Add new device information",
        "tags": ["DeviceInfo"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The device information was successfully added"
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "summary": "Get a user by userId",
        "tags": ["User"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DeviceInfo": {
        "type": "object",
        "required": ["userId", "baseOS", "version"],
        "properties": {
          "userId": {
            "type": "string",
            "description": "The user ID"
          },
          "baseOS": {
            "type": "string",
            "description": "The base operating system"
          },
          "version": {
            "type": "string",
            "description": "The version of the device"
          }
        }
      },
      "User": {
        "type": "object",
        "required": ["userId", "name", "email"],
        "properties": {
          "userId": {
            "type": "string",
            "description": "The user ID"
          },
          "name": {
            "type": "string",
            "description": "The name of the user"
          },
          "email": {
            "type": "string",
            "description": "The email of the user"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "DeviceInfo",
      "description": "Device information management"
    },
    {
      "name": "User",
      "description": "User management"
    }
  ]
}